pyproject.toml
[project]
name = "garmin-ingest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "garminconnect>=0.2.36",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
